Create Determine if the array has a negative element.
Write a function called isNegativeInArray that takes the arr array as an argument and returns true if the array has at least one negative element, and false otherwise. You must use the break statement in your solution.
Indication. Set some variable to false. Loop through the array, and if a negative element is encountered, set this variable to true and interrupt the loop.
Examples:
* function isNegativeInArray ([1, 2, 3, 9, 0]) must return false;
* function isNegativeInArray ([2, 1, -3, 4, 3]) must return true;
* function isNegativeInArray ([2, 1, -3, -4, 3]) must return true.