Create Does the string contain the given character?
Write a function named iSymbInString that takes str and a symb as arguments and returns true if there is at least one symb in the string and false otherwise.
You must use the break statement in your solution. The use of string methods is prohibited.
Indication. Loop through the line, and if the required character is encountered, assign a value to some variable and interrupt the loop.
Examples:
* function iSymbInString ("abcd", "b") must return true;
* the iSymbInString ("abcd", "f") function must return false.