Create Table of numbers: n lines, 5 numbers each
Write a function called numericalTable that takes n as an argument and returns a table of numbers from 1 to n. Each line contains 5 numbers, separated by a space. All lines except the last one end with a \ n newline character.
For example, for n = 4 the line should be obtained:
"1 1 1 1 1 \ n2 2 2 2 2 \ n3 3 3 3 3 \ n4 4 4 4 4"
When outputting to the console, such a line is displayed as a table:
When outputting to the console, such a line is displayed as a table:
"1 1 1 1 1
2 2 2 2 2
3 3 3 3 3
4 4 4 4 4 "
Note: When checking the solution in tests, the line feed character is not displayed.
Examples:
* numericalTable (1) should return: "1 1 1 1 1";
* numericalTable (3) function should return: "1 1 1 1 1 \ n2 2 2 2 2 \ n3 3 3 3 3";
* numericalTable (0) should return: "".