Create Is the product of two numbers a safe number?
Write a function checkingForfSafe that takes two values a and b as arguments, multiplies a by b, and checks if the product is safe. The function must return true if the product of the numbers is a safe number, and false otherwise.
Examples:
checkingForfSafe (4000, 2000) must return true;
checkingForfSafe (9007199254740991, 10) should return false;
checkingForfSafe (Infinity, 2) should return false.