Create Change the sign of the larger of two numbers
Write a function named changeMax that takes two numbers a and b as arguments, reverses the sign of the larger of the numbers, and returns that number. If the numbers are equal, return the message "The numbers are equal".
Example:
changeMax (9, 4) should return -9;
changeMax (1, 6) should return -6;
changeMax (0, 4) should return -4;
changeMax (-1, -3) should return 1;
changeMax (4, 4) should return "The numbers are equal".