Create The opposite value in the array.
You are given an array arr and some number num. Determine if the array contains the opposite of num, and if so, return its index. If there is no such element, return undefined.
Write a function called oppositeInArray that takes an array arr and a number num as arguments and returns the index of the opposite of num.
Example:
the function oppositeInArray ([1, -2, 3, 5], 2) must return 1 (since the element arr [1] = -2 is the opposite of the given number 2).
the function oppositeInArray ([100, 36, 1, 50], -50) should return 3 (since the element arr [3] = 50 is the opposite of the given number -50).
function oppositeInArray ([10, 3, 8, 11], 11) must return undefined (there is no element -11 in the array).