Create Which of the two points is closer to point a?
Which of the two points is closer to point a?
Three points a, b, c with different coordinates lie on the number axis. Determine which of the two points: b or c is closer to point a.
Write a function called nearPoint that takes three numbers a, b, c (coordinates of three points on the number axis) as arguments and returns:
line "b", if point b is closer to point a,
line "c", if point c is closer to point a,
"distances are equal" if the distances from points b and c to point are equal.
Examples:
nearest point function (9, 6, 14) should return "b";
closest point (3, 0, 4) should return "c";
closest point (0, 2, -4) should return "b".