Create Purchase payment
Claudia buys sweets at the store, her purchase is $ XX.XX. She is going to pay in cash, she has a bundle of dollar bills, and there are no coins. What is the minimum amount of dollars Claudia must give to the cashier in order to pay for the purchase?