Create Random phone number.
When testing the user registration page, you need to test the "Phone number" field. You have a lot of tests to do and you decide to create random phone numbers - 11-digit numbers. The phone number must start with a "+" character, the first digit of the number must be 1.

Write a function called createPhoneNumber that returns a random phone number as a 12-length string starting with '+1' and random numbers.