Create Average Traveler Speed
The traveler traveled s1 km from point A to point B in t1 hours, then s2 km from point B to point C in t2 hours. Find the average speed of the traveler and round the result to 2 decimal places.
Write a function averageSpeed that takes 4 numbers s1, t1, s2, t2 as arguments and returns the average speed of the traveler. Round the result to 2 decimal places and return it as a number.