Create The price of the goods on the check
The store's receipt contains the price of the product and the name of the product in the form: "2.59 milk". Get the price of an item as a number.

Write a function priceOfItem that takes a receipt (purchase description as a string: "price item name") as an argument and returns a number - the price of the item.