Create Arithmetic mean of array elements
Write a function average that takes an array arr as an argument and returns the average of the elements in that array. To find the arithmetic mean of the array elements, the sum of the array elements must be divided by the length of the array. If the array is empty (i.e., its length is 0), the function should return 0.