Create Array of temperatures in Fahrenheit
Write a function fahrenheit that takes two integers t1 and t2 (t1 <t2 is the starting and ending temperature in Celsius) and returns an array of Fahrenheit temperatures for a given temperature range. Temperatures in degrees Celsius and degrees Fahrenheit are related by the formula: tf = tc * (9/5) + 32 The calculated values are rounded to one decimal place.