Create Identical elements in an array
Write a function equalElementsInArray that takes an array of numbers arr as an argument and determines if the array has the same elements, and if so, returns them as an array [element, element]. If there are several such pairs, return the first one. If there are no identical elements in the array, return the message "All elements are different". If the original array is empty, return an empty array.