Create How many students have a grade above average?
The array contains the marks of the students in the class. How many students have a higher than average score? (That is, how many students have a score that is higher than the arithmetic mean of the array elements).
Write a function aboveAverageScore that takes an array of marks arr as an argument and returns the number of students who have an above average score. In all tests, the array is not empty.
Use a loop in your solution. Array methods are not allowed.