Create Swap the maximum and minimum elements of an array
Write a function changeMaxAndMin that takes the arr array as an argument and returns an array in which the maximum and minimum elements are swapped. If the array has several maximum or several minimum elements, swap the first of them. In all tests, the array contains at least two elements.

Use loops in your solution. The use of the Math.min () and Math.max () methods is not permitted.