Create The shortest and longest word in an array
Write a words function that takes an array of strings arr as an argument and returns the shortest and longest word as an array [shortestWord, longestWord]. It is guaranteed that in all tests the array contains at least one word. If the words are the same length, then the first in the list is returned.

Use loops in your solution. The use of the Math.min () and Math.max () methods is not permitted.