Create Maximum element modulus
Write a function maxAbsValue that takes an array arr as an argument and returns the element that has the maximum absolute value (modulus). If the values are equal, the first in the list is returned.

Use a loop in your solution. The use of the Math.max () method is not permitted.