Create Maximum even element of an array
Write a function maxEven that takes an array arr as an argument and returns the maximum even element of the array. If the array contains no even elements or the array is empty, return undefined.

Use a loop in your solution. The use of the Math.max () method is not permitted.