Create The coldest day of the month.
The specialists of the weather observation center record the air temperature every day and write the data into the array.
An array of 30 values ​​stores air temperature data in April.
For example, the array contains the values ​​[12, 13, 16, 10, ..., 20].
Let's assume the lowest temperature in this array is 10. This means that the coldest day was April 4th. (Remember that the numbering of the days of the month starts at 1, and the numbering of the elements in the array starts at zero).Determine which day was the coldest. If the lowest temperature was several days, return the last day.