Create Remove all non-numeric elements at the beginning of the array
An array is given, at the beginning of which there is a certain number of non-numeric elements, after which there may be numbers. Write a function deleteNotNumbers that takes a mixed array arr as an argument, removes all non-numeric array elements before the first number, and returns the resulting array.